refactor: interface, class 네이밍 수정
